backend/requirements.txt
flask
flask-cors
flask-sqlalchemy
pymysql
pyjwt
werkzeug
python-dotenv
